made some changed from beta testing
changed some descriptions to make the puzzles a little easier. moved
the npc to a different part of the castle so you have to explore more
of it. moved key to exit in the treasure room instead of in another
room because once you go down into the dungeon the door locks you in
there.